+++ packages/gradient/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.10.6"></a>
+## [0.10.6](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.10.5...@times-components/gradient@0.10.6) (2018-04-30)
+
+
+
+
+**Note:** Version bump only for package @times-components/gradient
+
 <a name="0.10.5"></a>
 ## [0.10.5](https://github.com/newsuk/times-components/compare/@times-components/gradient@0.10.4...@times-components/gradient@0.10.5) (2018-04-30)