+++ packages/related-articles/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.3.8"></a>
+## [0.3.8](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.3.7...@times-components/related-articles@0.3.8) (2018-04-30)
+
+
+
+
+**Note:** Version bump only for package @times-components/related-articles
+
 <a name="0.3.7"></a>
 ## [0.3.7](https://github.com/newsuk/times-components/compare/@times-components/related-articles@0.3.6...@times-components/related-articles@0.3.7) (2018-04-30)