+++ packages/video/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.2.4"></a>
+## [1.2.4](https://github.com/newsuk/times-components/compare/@times-components/video@1.2.3...@times-components/video@1.2.4) (2018-04-30)
+
+
+### Bug Fixes
+
+* video package code review updates ([#898](https://github.com/newsuk/times-components/issues/898)) ([fab3593](https://github.com/newsuk/times-components/commit/fab3593))
+
+
+
+
 <a name="1.2.3"></a>
 ## [1.2.3](https://github.com/newsuk/times-components/compare/@times-components/video@1.2.2...@times-components/video@1.2.3) (2018-04-30)